hands-on/archlens/pyproject.toml
[project]
name = "archlens"
version = "0.1.0"
description = "ArchLens: see engineering blogs the way architects do"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "feedparser>=6.0.12",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "sqlite-vec>=0.1.6",
    "streamlit>=1.31.0",
    "trafilatura>=2.0.0",
]
